Added a solution to exercise 4.6

+++ 4_6/src/main.c
@@ -4,9 +4,9 @@
 /* AFLEVERINGS OPGAVE 4.6 */
 
 // Funktion for at finde GCD
-int FindGcd(int num1, int num2) {
+int FindGcd(long num1, long num2) {
   while (num2 != 0) {
-    int temp = num2;
+    long temp = num2;
     num2 = num1 % num2;
     num1 = temp;
   }
